requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
astroid==3.3.9
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cpi==2.0.7
cryptography==45.0.4
cycler==0.12.1
dateparser==1.2.1
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
executing==2.2.0
fastapi==0.115.12
fonttools==4.58.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
narwhals==1.41.0
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.6
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pylint==3.3.7
pymongo==4.13.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pyzmq==26.4.0
redis==6.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
shap==0.47.2
shelved-cache==0.3.1
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.45.1
tabulate==0.8.10
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
wbdata==1.0.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
yellowbrick==1.5
